create new `reactpy.reactjs` module
