+++ src/reactpy/reactjs/types.py
@@ -1,0 +1,7 @@
+from reactpy.types import SourceType
+
+NAME_SOURCE = SourceType("NAME")
+"""A named source - usually a Javascript package name"""
+
+URL_SOURCE = SourceType("URL")
+"""A source loaded from a URL, usually a CDN"""
