+++ src/reactpy/web/__init__.py
@@ -3,13 +3,6 @@ from reactpy.web.module import (
     module_from_file,
     module_from_string,
     module_from_url,
-    reactjs_component_from_file,
-    reactjs_component_from_npm,
-    reactjs_component_from_string,
-    reactjs_component_from_url,
-)
-from reactpy.web.utils import (
-    reactjs_import_map,
 )
 
 __all__ = [
@@ -17,9 +10,4 @@ __all__ = [
     "module_from_file",
     "module_from_string",
     "module_from_url",
-    "reactjs_component_from_file",
-    "reactjs_component_from_npm",
-    "reactjs_component_from_string",
-    "reactjs_component_from_url",
-    "reactjs_import_map",
 ]
