+++ tests/test_reactjs/js_fixtures/export-resolution/index.js
@@ -1,0 +1,2 @@
+export { index as Index };
+export * from "./one.js";
