+++ tests/test_reactjs/js_fixtures/export-resolution/one.js
@@ -1,0 +1,5 @@
+export { one as One };
+// use ../ just to check that it works
+export * from "../export-resolution/two.js";
+// this default should not be exported by the * re-export in index.js
+export default 0;
