+++ tests/test_reactjs/js_fixtures/export-resolution/two.js
@@ -1,0 +1,2 @@
+export { two as Two };
+export * from "https://some.external.url";
