+++ tests/test_reactjs/js_fixtures/generic-module.js
@@ -1,0 +1,5 @@
+import { h } from "https://unpkg.com/preact?module";
+
+export function GenericComponent(props) {
+  return h("div", { id: props.id }, props.text);
+}
